# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Forge local control API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Local-only HTTP/JSON and SSE control contract for Forge.

    This API is served over an owner-only Unix domain socket, not a TCP
    listener. The nominal server URL only supplies an HTTP authority to generic
    OpenAPI clients. M0 has no application-level authentication: the local
    process is the sole operator. Every mutation is a named command, committed
    by Core, and requires an `Idempotency-Key` header plus an expected Project
    revision. The operator is recorded in canonical Events as a `human` actor;
    M0 defines no separate local-administrator actor kind.
servers:
  - url: http://forge.local
    description: Nominal HTTP authority carried over the local Unix socket.
...
security: []
...
paths:
...
  /v1/projects/{project_id}/memory:
    get:
      tags: [memory]
      operationId: listDerivedMemory
      summary: List current, visible derived memory entries.
      description: |
        Without employee_id, only Project-visible entries are returned. With a
        same-Project employee_id, that Employee's personal entries are included.
        Withdrawn entries and entries whose evidence is no longer valid are omitted.
        The cursor advances over scanned candidates, so a page can contain fewer
        items than limit even when next_cursor is present. Owner-only entry and
        history reads are separate audit routes and may include withdrawn content.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: employee_id, in: query, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: after, in: query, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
      responses:
        '200':
          description: Ascending entry-ID scan with a candidate cursor.
          content: {application/json: {schema: {$ref: '#/components/schemas/DerivedMemoryList'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/projects/{project_id}/memory/search:
    get:
      tags: [memory]
      operationId: searchMemory
      summary: Search scoped canonical knowledge and derived memory.
      description: |
        Search is bounded retrieval, not authority or an exhaustive listing.
        Indexed hits are rehydrated from canonical Forge records with scope and
        evidence rechecked. When the index is absent or unavailable, Core uses a
        bounded canonical substring fallback. The result reports mode, degradation
        and truncation; callers must not treat score or projection_id as canonical.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: query, in: query, required: true, schema: {type: string, minLength: 1, x-forge-max-utf8-bytes: 8192}}
        - {name: employee_id, in: query, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
      responses:
        '200':
          description: Up to limit results with an overall serialized hit budget of 256 KiB.
          content: {application/json: {schema: {$ref: '#/components/schemas/MemorySearchResult'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/projects/{project_id}/memory/status:
    get:
      tags: [memory]
      operationId: getMemoryProjectionStatus
      summary: Read Project indexing diagnostics; not an indexing acknowledgement.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
      responses:
        '200':
          description: Current projection counts and configuration state.
          content: {application/json: {schema: {$ref: '#/components/schemas/MemoryProjectionStatus'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/projects/{project_id}/memory/{id}:
    get:
      tags: [memory]
      operationId: getDerivedMemoryEntry
      summary: Read one current derived memory revision for owner audit.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
      responses:
        '200':
          description: Canonical head, which may be withdrawn or personal.
          content: {application/json: {schema: {$ref: '#/components/schemas/DerivedMemoryEntry'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/projects/{project_id}/memory/{id}/history:
    get:
      tags: [memory]
      operationId: listDerivedMemoryHistory
      summary: Read immutable revision history for owner audit.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: after, in: query, schema: {type: integer, minimum: 0, maximum: 9223372036854775807, format: int64}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
      responses:
        '200':
          description: Ascending revisions strictly after the supplied revision.
          content: {application/json: {schema: {$ref: '#/components/schemas/DerivedMemoryHistory'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/projects/{project_id}/employees/{employee_id}/onboarding:
    get:
      tags: [memory]
      operationId: getEmployeeOnboarding
      summary: Read one Employee's onboarding status in the Project.
      parameters:
        - {name: project_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
        - {name: employee_id, in: path, required: true, schema: {$ref: '#/components/schemas/UuidV7'}}
      responses:
        '200':
          description: Persisted onboarding state and optional job receipt.
          content: {application/json: {schema: {$ref: '#/components/schemas/EmployeeOnboardingStatus'}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
...
  responses:
    InvalidRequest:
      description: The request cannot be parsed or violates a transport contract.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The server-derived actor lacks the required capability.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource does not exist in the Project scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unavailable:
      description: A required local Forge dependency is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    UuidV7:
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      description: A UUIDv7 identifier generated by Forge unless stated otherwise.
    Rfc3339Timestamp:
      type: string
      format: date-time
      description: UTC timestamp serialized in RFC 3339 format.
    JsonObject:
      type: object
      additionalProperties: true
      description: A JSON object interpreted only by the selected domain contract.
...
    KnowledgeSourceRef:
      oneOf:
        - type: object
          additionalProperties: false
          required: [kind, artifact_id]
          properties:
            kind: {const: artifact}
            artifact_id: {$ref: '#/components/schemas/UuidV7'}
        - type: object
          additionalProperties: false
          required: [kind, event_id]
          properties:
            kind: {const: event}
            event_id: {$ref: '#/components/schemas/UuidV7'}
        - type: object
          additionalProperties: false
          required: [kind, handoff_id]
          properties:
            kind: {const: task_handoff}
            handoff_id: {$ref: '#/components/schemas/UuidV7'}
        - type: object
          additionalProperties: false
          required: [kind, page_id, revision]
          properties:
            kind: {const: knowledge_page}
            page_id: {$ref: '#/components/schemas/UuidV7'}
            revision: {type: integer, minimum: 1, format: int64}
    KnowledgePageContent:
      type: object
      additionalProperties: false
      required: [title, markdown, source_refs]
      properties:
        title: {type: string, minLength: 1, x-forge-max-utf8-bytes: 200}
        markdown: {type: string, minLength: 1, x-forge-max-utf8-bytes: 65536}
        source_refs:
          type: array
          maxItems: 128
          uniqueItems: true
          items: {$ref: '#/components/schemas/KnowledgeSourceRef'}
...
    KnowledgePage:
      type: object
      additionalProperties: false
      required: [id, project_id, revision, kind, status, content, content_hash, actor, command_id, created_at, revised_at, supersedes_revision]
      properties:
        id: {$ref: '#/components/schemas/UuidV7'}
        project_id: {$ref: '#/components/schemas/UuidV7'}
        revision: {type: integer, minimum: 1, format: int64}
        kind: {type: string, enum: [introduction, architecture, guide, policy, decision]}
        status: {type: string, enum: [draft, published, withdrawn]}
        content: {$ref: '#/components/schemas/KnowledgePageContent'}
        content_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
        actor: {$ref: '#/components/schemas/ActorReference'}
        command_id: {$ref: '#/components/schemas/UuidV7'}
        created_at: {$ref: '#/components/schemas/Rfc3339Timestamp'}
        revised_at: {$ref: '#/components/schemas/Rfc3339Timestamp'}
        supersedes_revision: {type: [integer, 'null'], minimum: 1, format: int64}
      description: Only published Policy and Decision pages carry managed authority.
...
    DerivedMemorySubject:
      oneOf:
        - type: object
          additionalProperties: false
          required: [kind, task_id]
          properties:
            kind: {const: task_summary}
            task_id: {$ref: '#/components/schemas/UuidV7'}
        - type: object
          additionalProperties: false
          required: [kind, employee_id, task_id]
          properties:
            kind: {const: employee_memory_entry}
            employee_id: {$ref: '#/components/schemas/UuidV7'}
            task_id: {type: [string, 'null'], format: uuid}
        - type: object
          additionalProperties: false
          required: [kind, task_id]
          properties:
            kind: {const: project_knowledge_entry}
            task_id: {type: [string, 'null'], format: uuid}
    DerivedMemoryEntry:
      type: object
      additionalProperties: false
      required: [id, project_id, revision, subject, markdown, source_refs, content_hash, coverage, created_by_job_id, created_at, withdrawn]
      properties:
        id: {$ref: '#/components/schemas/UuidV7'}
        project_id: {$ref: '#/components/schemas/UuidV7'}
        revision: {type: integer, minimum: 1, format: int64}
        subject: {$ref: '#/components/schemas/DerivedMemorySubject'}
        markdown: {type: string, minLength: 1, x-forge-max-utf8-bytes: 65536}
        source_refs: {type: array, minItems: 1, maxItems: 128, uniqueItems: true, items: {$ref: '#/components/schemas/KnowledgeSourceRef'}}
        content_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
        coverage:
          type: [object, 'null']
          additionalProperties: false
          required: [first_event_sequence, last_event_sequence]
          properties:
            first_event_sequence: {type: integer, minimum: 1, format: int64}
            last_event_sequence: {type: integer, minimum: 1, format: int64}
        created_by_job_id: {$ref: '#/components/schemas/UuidV7'}
        created_at: {$ref: '#/components/schemas/Rfc3339Timestamp'}
        withdrawn: {type: boolean}
      description: Evidence-backed generated content; never an authoritative Policy or Decision.
    DerivedMemoryList:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/DerivedMemoryEntry'}}
        next_cursor: {type: [string, 'null'], format: uuid}
    DerivedMemoryHistory:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/DerivedMemoryEntry'}}
        next_cursor: {type: [string, 'null'], pattern: '^[0-9]+$'}
    MemorySearchDocument:
      oneOf:
        - type: object
          additionalProperties: false
          required: [kind, record]
          properties:
            kind: {const: knowledge_page}
            record: {$ref: '#/components/schemas/KnowledgePage'}
        - type: object
          additionalProperties: false
          required: [kind, record]
          properties:
            kind: {const: derived_memory}
            record: {$ref: '#/components/schemas/DerivedMemoryEntry'}
    MemorySearchResult:
      type: object
      additionalProperties: false
      required: [project_id, employee_id, mode, degradation, truncated, results]
      properties:
        project_id: {$ref: '#/components/schemas/UuidV7'}
        employee_id: {type: [string, 'null'], format: uuid}
        mode: {type: string, enum: [indexed, canonical_fallback]}
        degradation: {type: [string, 'null'], enum: [not_configured, index_unavailable, rejected_projection, null]}
        truncated: {type: boolean}
        results:
          type: array
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required: [projection_id, score, document]
            properties:
              projection_id: {type: [string, 'null'], format: uuid}
              score: {type: [number, 'null']}
              document: {$ref: '#/components/schemas/MemorySearchDocument'}
    MemoryProjectionStatus:
      type: object
      additionalProperties: false
      required: [indexed, pending, retired, failed_attempts, canonical_max_revision, configured]
      properties:
        indexed: {type: integer, minimum: 0}
        pending: {type: integer, minimum: 0}
        retired: {type: integer, minimum: 0}
        failed_attempts: {type: integer, minimum: 0}
        canonical_max_revision: {type: integer, minimum: 0}
        configured: {type: boolean}
    EmployeeOnboardingStatus:
      type: object
      additionalProperties: false
      required: [employee_id, state, revision, job_id, receipt]
      properties:
        employee_id: {$ref: '#/components/schemas/UuidV7'}
        state: {type: string, enum: [pending, completed, skipped, legacy_bypass]}
        revision: {type: integer, minimum: 1, format: int64}
        job_id: {type: [string, 'null'], format: uuid}
        receipt: {type: [object, 'null'], additionalProperties: true}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - validation_failed
            - not_found
            - conflict
            - stale_revision
            - idempotency_conflict
            - forbidden
            - execution_stopped
            - cursor_invalid
            - unavailable
            - internal
        message:
          type: string
        details:
          $ref: '#/components/schemas/JsonObject'
        request_id:
          $ref: '#/components/schemas/UuidV7'
...
    ActorReference:
      type: object
      additionalProperties: false
      required: [kind, id]
      properties:
        kind:
          type: string
          enum: [human, employee, system_manager, core, supervisor]
        id:
          $ref: '#/components/schemas/UuidV7'